--- ARCHITECTURE_REFACTOR_SUMMARY.md
@@ -1,196 +1,0 @@
-# 新闻收集器架构重构总结
-
-## 🎯 重构目标
-
-将原本的单体爬虫架构重构为**两阶段解耦架构**，提高系统的灵活性和可维护性。
-
-## 📊 架构对比
-
-### 旧架构 (单体集成)
-```
-新闻网站 → 内置爬虫 → 直接创建Document → RAGFlow知识库
-```
-
-**问题：**
-- 爬虫逻辑与RAGFlow强耦合
-- 难以支持复杂网站结构
-- 维护困难，扩展性差
-- 无法利用专业爬虫工具
-
-### 新架构 (两阶段解耦)
-```
-阶段1: 新闻网站 → 外部爬虫工具 → 标准化本地目录
-阶段2: 本地目录 → RAGFlow文件夹上传 → 知识库文档
-```
-
-**优势：**
-- ✅ 关注点分离：爬虫专注抓取，RAGFlow专注处理
-- ✅ 工具专业化：可使用Scrapy、Selenium等专业工具
-- ✅ 标准化接口：统一的目录结构和元数据格式
-- ✅ 复用现有功能：利用已有的文件夹上传机制
-- ✅ 易于维护：爬虫工具可独立开发和升级
-
-## 🔧 实施变更
-
-### 1. 简化news_fetcher.py
-**变更前：**
-- 复杂的网站解析逻辑
-- BeautifulSoup依赖
-- 网站特定的选择器
-
-**变更后：**
-- 只保留任务执行器框架
-- 支持调用外部爬虫工具
-- 生成演示内容用于测试
-
-### 2. 修改集成服务
-**变更前：**
-- 直接创建Document和File记录
-- 内存中处理新闻数据
-
-**变更后：**
-- 调用外部爬虫生成文件
-- 使用文件夹上传机制
-- 标准化的元数据处理
-
-### 3. 扩展数据库模型
-**新增字段：**
-```python
-class NewsTask:
-    crawler_config = JSONField(null=True, default={
-        "type": "demo",           # 爬虫类型
-        "timeout": 300,           # 超时设置
-        "output_format": "markdown"  # 输出格式
-    })
-```
-
-## 📁 标准化目录结构
-
-外部爬虫工具应按以下格式输出：
-
-```
-{output_directory}/
-├── metadata.json              # 任务元数据
-├── sources/                   # 新闻源目录
-│   ├── 新浪科技/
-│   │   ├── source_info.json  # 源信息
-│   │   ├── 文章1.md          # Markdown格式文章
-│   │   ├── 文章2.md
-│   │   └── ...
-│   ├── 网易科技/
-│   │   ├── source_info.json
-│   │   ├── 文章1.md
-│   │   └── ...
-│   └── 36氪/
-│       └── ...
-└── logs/                      # 日志目录
-    ├── crawler.log
-    └── errors.log
-```
-
-### 文章格式示例
-```markdown
----
-title: "人工智能技术突破新进展"
-url: "https://tech.sina.com.cn/article/123456"
-author: "张三"
-publish_time: "2025-01-26T09:30:00Z"
-source: "新浪科技"
-category: "人工智能"
-tags: ["AI", "技术", "突破"]
-crawled_at: "2025-01-26T10:05:30Z"
----
-
-# 人工智能技术突破新进展
-
-[文章正文内容...]
-```
-
-## 🔌 外部爬虫工具集成
-
-### 支持的集成方式
-
-1. **命令行调用**
-```python
-subprocess.run([
-    "python", "crawler_tool.py",
-    "--config", "/path/to/config.json",
-    "--output", "/tmp/news_crawl_20250126"
-])
-```
-
-2. **HTTP API调用**
-```python
-requests.post("http://crawler-service:8080/crawl", json={
-    "sources": [...],
-    "output_dir": "/tmp/news_crawl_20250126"
-})
-```
-
-3. **Python包导入**
-```python
-from news_crawler import NewsCrawler
-crawler = NewsCrawler(config)
-crawler.crawl(sources, output_dir)
-```
-
-### 推荐爬虫工具
-
-- **Scrapy**: 强大的爬虫框架，适合大规模抓取
-- **Newspaper3k**: 专门的新闻抓取库，智能内容提取
-- **Selenium**: 处理JavaScript渲染的动态页面
-- **Beautiful Soup**: 简单的HTML解析，适合轻量级需求
-
-## 🧪 测试验证
-
-### 当前测试能力
-- ✅ 演示爬虫：生成标准格式的测试文件
-- ✅ 文件夹集成：自动导入到RAGFlow知识库
-- ✅ 文档解析：支持自动解析和索引
-- ✅ 错误处理：完整的异常处理机制
-
-### 测试脚本
-- `news_architecture_test.py`: 验证两阶段架构的完整流程
-- 测试覆盖：创建源 → 创建任务 → 执行抓取 → 验证结果
-
-## 🚀 部署建议
-
-### 开发阶段
-1. 使用内置的演示爬虫验证流程
-2. 开发和测试外部爬虫工具
-3. 验证文件格式和元数据规范
-
-### 生产部署
-1. 部署专业爬虫工具（Scrapy集群等）
-2. 配置定时任务调度
-3. 监控爬虫执行和错误处理
-4. 设置日志收集和告警
-
-## 📈 扩展计划
-
-### 短期目标
-- ✅ 完成基础架构重构
-- ⏳ 集成一个实际的爬虫工具
-- ⏳ 完善错误处理和重试机制
-
-### 中期目标
-- 支持增量抓取和去重
-- 添加内容质量评估
-- 实现分布式爬虫支持
-
-### 长期目标
-- AI驱动的内容筛选
-- 多媒体内容支持
-- 实时新闻流处理
-
-## 💡 最佳实践
-
-1. **爬虫工具选择**：根据网站特性选择合适的工具
-2. **内容质量**：设置合理的过滤条件
-3. **频率控制**：避免对网站造成过大压力
-4. **错误处理**：完善的重试和降级机制
-5. **监控告警**：及时发现和处理问题
-
----
-
-**总结**：新的两阶段架构通过解耦爬虫和处理逻辑，大幅提升了系统的灵活性和可维护性。现在可以利用专业的爬虫工具处理复杂的网站结构，同时复用RAGFlow现有的文档处理能力，实现了最佳的关注点分离。
