--- NEWS_COLLECTOR_SUMMARY.md
@@ -1,120 +1,0 @@
-# 新闻收集器CRUD功能完成总结
-
-## 📋 功能概述
-新闻收集器模块现已完成所有CRUD操作的实现，包含完整的新闻源管理、新闻任务管理和新闻内容管理功能。
-
-## ✅ 已完成功能
-
-### 1. 新闻源管理 (News Sources)
-- **创建新闻源** - POST `/api/v1/news_collector/sources`
-- **获取新闻源列表** - GET `/api/v1/news_collector/sources`
-- **获取单个新闻源** - GET `/api/v1/news_collector/sources/{source_id}`
-- **更新新闻源** - PUT `/api/v1/news_collector/sources/{source_id}`
-- **删除新闻源** - DELETE `/api/v1/news_collector/sources/{source_id}`
-
-### 2. 新闻任务管理 (News Tasks)
-- **创建新闻任务** - POST `/api/v1/news_collector/tasks`
-- **获取任务列表** - GET `/api/v1/news_collector/tasks`
-- **获取单个任务** - GET `/api/v1/news_collector/tasks/{task_id}`
-- **更新任务** - PUT `/api/v1/news_collector/tasks/{task_id}`
-- **删除任务** - DELETE `/api/v1/news_collector/tasks/{task_id}`
-- **执行任务** - POST `/api/v1/news_collector/tasks/{task_id}/execute`
-
-### 3. 新闻内容管理 (News Content)
-- **获取新闻内容列表** - GET `/api/v1/news_collector/contents`
-- **获取单个新闻内容** - GET `/api/v1/news_collector/contents/{content_id}`
-- **删除新闻内容** - DELETE `/api/v1/news_collector/contents/{content_id}`
-
-### 4. 统计功能
-- **获取统计信息** - GET `/api/v1/news_collector/statistics`
-- **获取任务状态** - GET `/api/v1/news_collector/tasks/{task_id}/status`
-
-### 5. 高级功能
-- **新闻收集与解析** - 集成新闻爬虫和内容解析
-- **知识库集成** - 自动将解析的新闻内容导入知识库
-- **状态管理** - 完整的任务状态跟踪
-- **配置管理** - 灵活的爬虫配置选项
-
-## 🔧 技术修复记录
-
-### 类型安全问题修复
-- ✅ 修复所有 `BeartypeCallHintParamViolation` 错误
-- ✅ 正确使用 `Optional[str]` 类型注解
-- ✅ 修复 `AttributeError: 'super' object has no attribute 'to_dict'` 错误
-- ✅ 优化数据库查询和对象处理
-
-### 数据库模型优化
-- ✅ NewsSource、NewsTask、NewsContent 模型完善
-- ✅ 支持 user_id 为空的情况（自动使用 tenant_id）
-- ✅ 完整的外键关系和约束
-
-### API层改进
-- ✅ 遵循 RAGFlow SDK 规范
-- ✅ 使用 @manager.route 装饰器
-- ✅ 集成 @token_required 认证
-- ✅ 统一的错误处理和响应格式
-
-## 🧪 验证状态
-
-### 已通过测试
-- ✅ **类型安全测试** - 所有参数类型注解正确
-- ✅ **数据库CRUD测试** - 创建、查询、更新、删除操作正常
-- ✅ **参数处理测试** - 支持 None 值和字符串混合使用
-- ✅ **Beartype兼容性测试** - 完全兼容类型检查器
-
-### 测试覆盖范围
-- 基础CRUD操作 ✅
-- 参数验证和类型安全 ✅
-- 错误处理 ✅
-- 数据库约束 ✅
-- API认证和权限 ✅
-
-## 📁 文件结构
-
-### 核心文件
-- `api/apps/sdk/news_collector.py` - 主要API端点（20个接口）
-- `api/db/services/news_service.py` - 数据服务层
-- `api/db/db_models.py` - 数据模型定义
-
-### 文档文件
-- `NEWS_COLLECTOR_DEPLOYMENT_GUIDE.md` - 部署指南
-- `NEWS_COLLECTOR_QUICKSTART.md` - 快速开始
-- `NEWS_MODULE_SUMMARY.md` - 模块总结
-- `SDK_ENHANCED_GUIDE.md` - SDK使用指南
-
-## 🚀 使用方式
-
-### 启动服务
-确保RAGFlow主服务正在运行，新闻收集器API将自动可用。
-
-### API调用示例
-```bash
-# 创建新闻源
-curl -X POST http://localhost:9222/api/v1/news_collector/sources \
-  -H "Authorization: Bearer YOUR_TOKEN" \
-  -H "Content-Type: application/json" \
-  -d '{
-    "name": "示例新闻源",
-    "url": "https://example.com/news",
-    "status": "active"
-  }'
-
-# 获取新闻源列表
-curl -X GET http://localhost:9222/api/v1/news_collector/sources \
-  -H "Authorization: Bearer YOUR_TOKEN"
-```
-
-## 📝 注意事项
-
-1. **认证要求** - 所有API调用都需要有效的Bearer Token
-2. **权限控制** - 基于tenant_id的多租户隔离
-3. **参数类型** - 支持Optional参数，None值会被正确处理
-4. **错误处理** - 统一的错误响应格式
-
-## 🎯 完成状态
-
-**新闻收集器CRUD功能已完全实现并通过所有测试验证。现在可以在生产环境中安全使用。**
-
----
-*最后更新: 2025年7月27日*
-*状态: 生产就绪 ✅*
