+++ api/apps/sdk/crawlers/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+爬虫模块
+"""
+
+from .base_crawler import LibraryCrawler
+from .topic_crawler import TopicCrawler
+from .url_seeding_crawler import UrlSeedingCrawler
+
+__all__ = [
+    "LibraryCrawler",
+    "TopicCrawler",
+    "UrlSeedingCrawler",
+]
